requirements.txt
adal==1.2.2
attrs==19.3.0
azure-common==1.1.24
azure-mgmt-compute==4.6.2
azure-mgmt-network==2.7.0
azure-mgmt-resource==2.2.0
azure-mgmt-storage==7.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
easyimap==0.6.3
flanker==0.9.11
haikunator==2.1.0
idna==2.8
isodate==0.6.0
msrest==0.6.10
msrestazure==0.6.2
oauthlib==3.1.0
ply==3.11
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
regex==2019.12.20
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
tld==0.11.9
typing==3.7.4.1
urllib3==1.25.7
WebOb==1.8.5